requirements.txt
cssselect==0.9.1
future==0.15.2
lxml==3.6.0
pyquery==1.2.13
python-telegram-bot==3.4
requests==2.9.1
wheel==0.24.0
